fix(proxy): correct max body bytes error messages
- Updated error messages in `proxy.go` and `config.go` to explicitly cast `math.MaxInt64` to `int64`, ensuring clarity in error reporting for maximum body size constraints.
- This change improves the accuracy of error messages related to body size limits in the proxy configuration and validation processes.

+++ internal/ncmctl/proxy.go
@@ -138,7 +138,7 @@ func (c *Proxy) validate() error {
 	}
 	// Capture helpers reserve one extra byte to distinguish truncation.
 	if c.opts.MaxBodyBytes == math.MaxInt64 {
-		return fmt.Errorf("max-body must be less than %d bytes", math.MaxInt64)
+		return fmt.Errorf("max-body must be less than %d bytes", int64(math.MaxInt64))
 	}
 
 	seeds, err := c.xeapiSessionSeeds()
